requirements.txt
# LangChain and LangGraph
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# Vector Search
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
joblib>=1.4.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
